requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
crewai==1.6.1
cryptography==46.0.3
cycler==0.12.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
googleapis-common-protos==1.72.0
groq==0.36.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.1.6
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==7.2.1
importlib_resources==6.5.2
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.3.2
json5==0.12.1
json_repair==0.25.2
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.8.2
mcp==1.22.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
narwhals==2.12.0
nodeenv==1.9.1
numba==0.62.1
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.8.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==23.2
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==10.4.0
platformdirs==4.5.0
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.0
propcache==0.4.1
protobuf==6.33.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pypdfium2==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.29.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.16.3
seaborn==0.13.2
shap==0.50.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
streamlit==1.31.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tomli_w==1.2.0
tornado==6.5.2
tqdm==4.67.1
ty==0.0.1a29
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uv==0.9.13
uvicorn==0.38.0
validators==0.35.0
virtualenv==20.35.4
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
